requirements.txt
python_wordpress_xmlrpc
pymysql
python_dateutil
requests
